feat(#284): rename one class to ClassName -> ClassNameVisitor
